Do not render special pages (eg ConfigSideBar) for some URL routes.
ConfigSideBar, ConfigHeader, ConfigFooter special pages should not be
rendered for requests of the following types:

* HTTP Method is POST, unless it's a request for the edit page.
* The request is for a resource that doesn't exist.
* The request is for an attachment (any ctype parameter).
* The request is for a history page and the ctype=atom parameter is
  specified. NB: The pages should be rendered if any other ctype
  parameter is specified (ctype=raw, ctype=default).
* The request is for a wiki page and the ctype=raw parameter is
  specified. NB: The pages should be rendered if any other ctype
  parameter is specified (ctype=atom, ctype=default).

The logic that distinguishes between these cases is already present in
PageHandlerImpl. So we move the call to addSideBarEtcToRequest() from
WikiHandlerImpl to PageHandlerImpl.

https://jira.int.corefiling.com/browse/REVIKI-642
